# Added pre-processed preferred affinity terms to scheduler's PodInfo type