**What type of PR is this?**

 /kind feature

**What this PR does / why we need it**:
This is the second part of adding processed affinity terms to PodInfo, this PR does that for preferred terms. Roughly 25% performance improvement for inter-pod (anti)affinity score.

before
```
BenchmarkScheduling/5000Nodes/5000Pods-12                          1000  1411689   ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12           1000  8155520   ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/5000Pods-12               1000  12699040  ns/op
BenchmarkSchedulingPreferredPodAffinity/5000Nodes/5000Pods-12      1000  11805901  ns/op
BenchmarkSchedulingPreferredPodAntiAffinity/5000Nodes/5000Pods-12  1000  12808495  ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/5000Pods-12              1000  2725989   ns/op
```

after
```
BenchmarkScheduling/5000Nodes/5000Pods-12                          1000  1400054   ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12           1000  7228791   ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/5000Pods-12               1000  11914596  ns/op
BenchmarkSchedulingPreferredPodAffinity/5000Nodes/5000Pods-12      1000  9423096   ns/op
BenchmarkSchedulingPreferredPodAntiAffinity/5000Nodes/5000Pods-12  1000  10320671  ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/5000Pods-12              1000  2632871   ns/op
```


**Which issue(s) this PR fixes**:
Part of #91058

**Special notes for your reviewer**:
Builds on #91062

**Does this PR introduce a user-facing change?**:
```release-note
NONE
```